修改异步写入日志方法和增加获取绝对路径方法

+++ Logger/logger.go
@@ -7,6 +7,7 @@ import (
 	"os"
 	"path/filepath"
 	"runtime"
+	"strings"
 	"sync"
 	"time"
 )
@@ -16,6 +17,7 @@ type Logger interface {
 	Errorf(format string, a ...interface{})
 	Infof(format string, a ...interface{})
 	GetConf()
+	Close()
 }
 
 const (
@@ -25,12 +27,13 @@ const (
 )
 
 type Log struct {
-	LogLevel    int        // 日志级别
-	FilePath    string     // 文件存储路径
-	MaxDay      int64      // 最大存储天数
-	currentFile *os.File   // 当前文件
-	currentDate string     // 文件创建时的日期
-	mutex       sync.Mutex // 互斥锁
+	LogLevel    int         // 日志级别
+	FilePath    string      // 文件存储路径
+	MaxDay      int64       // 最大存储天数
+	currentFile *os.File    // 当前文件
+	currentDate string      // 文件创建时的日期
+	mutex       sync.Mutex  // 互斥锁
+	logChannels chan string // 异步写入
 }
 
 func NewLogger() Logger {
@@ -42,6 +45,7 @@ func (l *Log) InitLogger() {
 	l.LogLevel = Info
 	l.MaxDay = 7
 	l.FilePath = "."
+	l.logChannels = make(chan string, 3000)
 	// 清理日志文件
 	go func() {
 		err := l.clearOldLogs()
@@ -71,32 +75,43 @@ func (l *Log) SetLogger(Level int, FilePath string, MaxDay int64) {
 		}
 		l.FilePath = FilePath
 	}
+	l.FilePath = relativePathToAbsPath(l.FilePath)
 	l.MaxDay = MaxDay
 	FileName := formatLogFileName(time.Now())
 	File, err := os.OpenFile(l.FilePath+"/"+FileName, os.O_CREATE|os.O_APPEND|os.O_WRONLY, 0666)
 	if err != nil {
 		log.Fatal(err)
 	}
+
 	l.currentFile = File
 	l.currentDate = formatLogFileName(time.Now())
+	go l.logWriteToFile()
 }
 
-func (l *Log) logWriteToFile(format string, a ...interface{}) {
-	currentDate := time.Now().Format("2006-01-02")
-	if currentDate != l.currentDate {
-		l.createLogFile(time.Now())
-	}
-	Content := l.logWithCallerInfo(format, a...)
-	_, _ = l.currentFile.WriteString(Content)
+func (l *Log) logWriteToFile() {
+	for logline := range l.logChannels {
+		if logline != "" {
+			fmt.Println(logline)
+			currentDate := time.Now().Format("2006-01-02")
+			if currentDate != l.currentDate {
+				l.createLogFile(time.Now())
+			}
+			_, _ = l.currentFile.WriteString(logline)
 
-	// 检查并执行清理操作
-	go func() {
-		err := l.clearOldLogs()
-		if err != nil {
-			log.Println("Failed to clean old logs:", err)
+			// 检查并执行清理操作
+			go func() {
+				err := l.clearOldLogs()
+				if err != nil {
+					log.Println("Failed to clean old logs:", err)
+				}
+			}()
 		}
-	}()
+	}
+}
 
+func (l *Log) syncWriteLog(format string, a ...interface{}) {
+	message := l.logWithCallerInfo(fmt.Sprintf(format, a...))
+	l.logChannels <- message
 }
 
 func (l *Log) createLogFile(date time.Time) {
@@ -122,7 +137,7 @@ func (l *Log) Errorf(format string, a ...interface{}) {
 }
 
 func (l *Log) Infof(format string, a ...interface{}) {
-	l.logWriteToFile(format, a...)
+	l.syncWriteLog(format, a...)
 }
 
 func (l *Log) GetLevelString() string {
@@ -156,12 +171,11 @@ func formatLogFileName(data time.Time) string {
 }
 
 // 获取对应文件名，行号，方法名
-func (l *Log) logWithCallerInfo(format string, a ...interface{}) string {
+func (l *Log) logWithCallerInfo(logline string) string {
 	pc, file, line, _ := runtime.Caller(3)
 	funcName := runtime.FuncForPC(pc).Name()
-	message := fmt.Sprintf(format, a...)
 	Level := l.GetLevelString()
-	return fmt.Sprintf("[%s] fileLine:%s:%d funcName:%s;message:%s\n", Level, file, line, getFunctionName(funcName), message)
+	return fmt.Sprintf("[%s] fileLine:%s:%d funcName:%s;message:%s\n", Level, file, line, getFunctionName(funcName), logline)
 }
 
 // 获取对应的方法名
@@ -173,6 +187,9 @@ func getFunctionName(fullName string) string {
 			lastDotIndex = i
 			break
 		}
+	}
+	if fullName == "" {
+		return ""
 	}
 	return fullName[lastDotIndex+1:]
 }
@@ -195,8 +212,10 @@ func (l *Log) clearOldLogs() error {
 		// 检查文件日期是否早于需要清除的日期范围
 		if info.ModTime().Before(cutoffDate) {
 			// 删除文件
-			if err = os.Remove(path); err != nil {
-				return err
+			if strings.HasSuffix(path, ".log") {
+				if err = os.Remove(path); err != nil {
+					return err
+				}
 			}
 			log.Printf("Removed log file: %s\n", path)
 		}
@@ -207,4 +226,22 @@ func (l *Log) clearOldLogs() error {
 		return fmt.Errorf("failed to clear old logs:%v", err)
 	}
 	return nil
+}
+
+func relativePathToAbsPath(Path string) string {
+	absolutePath, err := filepath.Abs(Path)
+	if err != nil {
+		fmt.Println("Failed to get absolute path:", err)
+		return ""
+	}
+	fmt.Println(absolutePath)
+	return absolutePath
+}
+
+// 关闭对应的写入通道和文件
+func (l *Log) Close() {
+	close(l.logChannels)
+	if l.currentFile != nil {
+		_ = l.currentFile.Close()
+	}
 }
